+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 2.2.0 ()
 
 - [new] Statamic 4 support (with updated button icons)
+- [new] Option to show style name in the style selection button
 - [new] Toggle attribute values can now be configured
 - [new] Select type for attributes
-- []
-- []
+- [new] Automatic merging of multiple class/style based attributes
 - []
 - []
 - []